Refactors exchanges architecture, adds proxy
A new structure has been added: Proxy. Contains the necessary to intermediate between the websockets and the aggregator.
Common functions in each exchanges code (having a link with websockets) have been exported to a new package which contains the Proxy structure.

Finally, a bug have been discovered just after the last commit: when the websockets have nothing to send for a while, they close the connection.
A function have been added to resolve this bug. It recreates a new connection between the aggregator and the websocket.