Manage the subscriptions of each exchanges
We were not able to manage the unsubscriptions because we needed to have some informations about the current subscriptions (ex: chanId in Bitfinex).
To overcome this, we added the management of feedback messages sent by the exchangers during a modification in subscriptions.

The recoverer function presents in the proxy package has been refactored. It uses now the recover function provided by golang.